🔵 untangling assertions from tests
