+++ cmd/test/assert/validate_format_test.go
@@ -1,4 +1,4 @@
-package test
+package assert
 
 import (
 	"context"
